StartUp/requirements.txt
aiofiles==22.1.0
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
braintree==4.18.1
bs4==0.0.1
cachetools==4.2.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==39.0.2
csrf==0.1b1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.6
django-allauth==0.52.0
django-crispy-forms==2.0
django-filter==22.1
django-jazzmin==2.6.0
django-localflavor==3.1
django-location-field==2.1.0
django-parler==2.3
django-phonenumber-field==7.0.2
django-rosetta==0.9.8
django-seed==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.5
environs==9.5.0
exceptiongroup==1.1.1
fake-useragent==0.1.11
Faker==17.6.0
flower==1.2.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
humanize==4.6.0
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
loader==2017.9.11
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.3
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
packaging==22.0
pandas==1.5.2
phonenumbers==8.13.7
phonenumberslite==8.13.7
Pillow==9.3.0
polib==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-stdnum==1.18
python-telegram-bot==13.15
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.2
render==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
selenium==4.8.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telebot==0.0.5
telegram-send==0.34
toml==0.10.2
toposort==1.10
tornado==6.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
